pyproject.toml
[tool.poetry]
name = "examen"
version = "0.1.0"
description = "examen integritas"
authors = ["fl4r30n <flareon_rojo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
pymupdf = "^1.24.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
